feat: grava chamadas em MP3, volume tmpfs, retenção de 1h
Fases 1B/C/D do MVP (backlog em /home/dev/.claude/plans/frolicking-fluttering-puppy.md):

- B: ffmpeg no Dockerfile; audio_uploader.py converte tx.raw/rx.raw para
  tx.mp3/rx.mp3 (PCM16 8kHz mono -> libmp3lame) após escrever o bruto,
  apaga o .raw só se a conversão for bem-sucedida (se falhar, mantém o
  .raw e reporta status "uploaded_raw_only" em vez de perder o áudio).
  Validado: conversão real gera mp3 tocável, .raw é removido.

- C: docker-compose.app.yml troca o bind mount ./data/recordings pelo
  named volume recordings_tmpfs (driver tmpfs, 512MB) — gravação em RAM,
  não no HD do sistema. Montado em fastapi-1/2, arq-uploader, arq-cleanup.
  Validado: `mount` dentro do container confirma tmpfs real.

- D: AUDIO_RETENTION_DAYS vira float (permite ~1h = 0.0417); cron de
  limpeza passa de 1x/dia (03:00) para a cada 15 min.

Dois bugs reais pré-existentes encontrados e corrigidos ao subir
arq-cleanup pela primeira vez (nunca tinha rodado com sucesso):
cron(..., run_on_startup=...) usava kwarg errado (arq espera
run_at_startup), e WorkerSettings.redis_settings recebia a DSN como
string crua em vez de RedisSettings.from_dsn(...) — mesmo padrão já
usado corretamente em audio_uploader.py.

+++ src/config.py
@@ -29,7 +29,7 @@ class Settings(BaseSettings):
     STT_FALLBACK_TIMEOUT_MS: int = 500
     BATCH_INSERT_INTERVAL_SECONDS: int = 5
 
-    AUDIO_RETENTION_DAYS: int = 90
+    AUDIO_RETENTION_DAYS: float = 90
     RECORDINGS_PATH: str = "/data/recordings"
 
     model_config = {"env_file": ".env", "env_file_encoding": "utf-8"}
